restore RevTree.size() property to quicky obtain the total number of features in a tree. The property value is computed by RevTreeBuilder and part of the immutable state of the tree once built.

+++ src/cli/src/main/java/org/geogit/cli/porcelain/Add.java
@@ -74,7 +74,7 @@ public class Add extends AbstractCommand implements CLICommand {
         }
 
         console.print("Counting unstaged features...");
-        int unstaged = cli.getGeogit().getRepository().getWorkingTree().countUnstaged(pathFilter);
+        long unstaged = cli.getGeogit().getRepository().getWorkingTree().countUnstaged(pathFilter);
         if (0 == unstaged) {
             console.println();
             console.println("No unstaged features, exiting.");
@@ -92,7 +92,7 @@ public class Add extends AbstractCommand implements CLICommand {
         }
         WorkingTree workTree = op.setProgressListener(cli.getProgressListener()).call();
 
-        int staged = cli.getGeogit().getRepository().getIndex().countStaged(null);
+        long staged = cli.getGeogit().getRepository().getIndex().countStaged(null);
         unstaged = workTree.countUnstaged(null);
 
         console.println(staged + " features staged for commit");
